refactor: [#1221] move whitelist manager setup to whitelist context
