+++ src/bootstrap/app.rs
@@ -29,13 +29,13 @@ use crate::core::authentication::key::repository::persisted::DatabaseKeyReposito
 use crate::core::authentication::service;
 use crate::core::databases::setup::initialize_database;
 use crate::core::scrape_handler::ScrapeHandler;
-use crate::core::services::initialize_whitelist_manager;
 use crate::core::statistics;
 use crate::core::torrent::manager::TorrentsManager;
 use crate::core::torrent::repository::in_memory::InMemoryTorrentRepository;
 use crate::core::torrent::repository::persisted::DatabasePersistentTorrentRepository;
 use crate::core::whitelist::authorization::WhitelistAuthorization;
 use crate::core::whitelist::repository::in_memory::InMemoryWhitelist;
+use crate::core::whitelist::setup::initialize_whitelist_manager;
 use crate::servers::udp::server::banning::BanService;
 use crate::servers::udp::server::launcher::MAX_CONNECTION_ID_ERRORS_PER_IP;
 use crate::shared::crypto::ephemeral_instance_keys;
